+++ planning/autoware_trajectory_processor/include/autoware/trajectory_processor/trajectory_processor_data.hpp
@@ -17,15 +17,20 @@
 
 #include "autoware/trajectory_processor/semantic_speed_tracker.hpp"
 
+#include <autoware_map_msgs/msg/lanelet_map_bin.hpp>
 #include <autoware_perception_msgs/msg/predicted_objects.hpp>
+#include <autoware_perception_msgs/msg/tracked_objects.hpp>
 #include <autoware_perception_msgs/msg/traffic_light_group_array.hpp>
 #include <autoware_planning_msgs/msg/lanelet_route.hpp>
+#include <autoware_vehicle_msgs/msg/steering_report.hpp>
 #include <geometry_msgs/msg/accel_with_covariance_stamped.hpp>
 #include <nav_msgs/msg/odometry.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
+#include <std_msgs/msg/header.hpp>
 
 #include <lanelet2_core/LaneletMap.h>
 
+#include <cstddef>
 #include <memory>
 
 namespace autoware::trajectory_processor
@@ -39,12 +44,18 @@ struct TrajectoryProcessorData
 {
   nav_msgs::msg::Odometry::ConstSharedPtr current_odometry{nullptr};
   geometry_msgs::msg::AccelWithCovarianceStamped::ConstSharedPtr current_acceleration{nullptr};
+  autoware_vehicle_msgs::msg::SteeringReport::ConstSharedPtr current_steering{nullptr};
   autoware_perception_msgs::msg::PredictedObjects::ConstSharedPtr predicted_objects{nullptr};
+  autoware_perception_msgs::msg::TrackedObjects::ConstSharedPtr tracked_objects{nullptr};
   sensor_msgs::msg::PointCloud2::ConstSharedPtr obstacle_pointcloud{nullptr};
   std::shared_ptr<lanelet::LaneletMap> lanelet_map{nullptr};
+  autoware_map_msgs::msg::LaneletMapBin::ConstSharedPtr lanelet_map_bin{nullptr};
   autoware_planning_msgs::msg::LaneletRoute::ConstSharedPtr route{nullptr};
   autoware_perception_msgs::msg::TrafficLightGroupArray::ConstSharedPtr traffic_light_signals{
     nullptr};
+  std_msgs::msg::Header candidate_header{};
+  std::size_t candidate_index{0U};
+  std::size_t candidate_count{0U};
   SemanticSpeedTracker semantic_speed_tracker;
 };